Fixed some rounding errors, added CP adjusted results

+++ src/BattleMatch/Lookup/IPokemonTemplates.cs
@@ -7,9 +7,9 @@ namespace Optimizer.Lookup
     {
         PokemonTemplate GetTemplate(string pokemonName);
 
-        int GetMaxHP();
+        double GetMaxHP();
 
-        int GetMaxCP();
+        double GetMaxCP();
 
         IEnumerable<string> GetAllNames();
